Move transforms into texture module
